🚚Rename: 검색 엔진, 캐러셀 에셋 폴더 분리
