+++ src/components/main/BannerCarousel.jsx
@@ -3,8 +3,8 @@ import styled from 'styled-components';
 import Slider from 'react-slick';
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
-import currentDotImg from '../../assets/main/bannerDotCurrent.svg';
-import defaultDotImg from '../../assets/main/bannerDotDefault.svg';
+import currentDotImg from '../../assets/carousel/bannerDotCurrent.svg';
+import defaultDotImg from '../../assets/carousel/bannerDotDefault.svg';
 import Banner from './Banner';
 import LeftArrow from '../common/button/LeftArrow';
 import RightArrow from '../common/button/RightArrow';
