+++ src/components/main/PopularRecreationCarousel.jsx
@@ -5,8 +5,8 @@ import Slider from 'react-slick';
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
 import RecreationTripleSet from './RecreationTripleSet';
-import wholeSlide from '../../assets/main/wholeSlide.svg';
-import currentSlide from '../../assets/main/currentSlide.svg';
+import wholeSlide from '../../assets/carousel/wholeSlide.svg';
+import currentSlide from '../../assets/carousel/currentSlide.svg';
 import useLoginStore from '../../stores/loginStore';
 import LeftArrow from '../common/button/LeftArrow';
 import RightArrow from '../common/button/RightArrow';
